requirements.txt
tensorflow==1.9.0
numpy==1.14.5
opencv_contrib_python==3.4.1.15
progressbar2==3.39.2